pyproject.toml
[tool.poetry]
name = "pythoncontactsapi"
version = "0.1.0"
description = ""
authors = ["IrinaShushkevych <iraolijnyk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
SQLAlchemy = "^2.0.1"
psycopg2-binary = "^2.9.5"
configparser = "^5.3.0"
alembic = "^1.9.2"
libgravatar = "^1.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
fastapi-mail = "^1.2.5"
python-dotenv = "^1.0.0"
redis = "^4.5.1"
cloudinary = "^1.32.0"
fastapi-limiter = "^0.1.5"

[tool.poetry.dev-dependencies]
Sphinx = "^6.1.3"
pytest = "^7.2.2"
httpx = "^0.23.3"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
